clean up media-encoder worker
