+++ media-encoder/media-encoder.yaml
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 runtime: nodejs10
-service: encoding
+service: media-encoder
 instance_class: B2
 
 basic_scaling:
